third_party/blink/web_tests/platform/generic/fast/lists/ol-nested-list-expected.txt

This tests that list items are properly marked when the list contains nested lists.

10 Ten
11 Eleven
12 Twelve
13 Thirteen One Two
-> -> ◦ One
-> -> ◦ Two
14 Fourteen
15 Fifteen
-> 1 One
-> 2 Two
16 Sixteen
17 Seventeen
